Set ODBC time option in configure_connection hook
Recent changes to the ActiveRecord::ConnectionAdapters logic introduced
an explicit `configure_connection` hook for setting up any configuration
that might be necessary for a newly established database connection.
Previously, we'd relied on the one custom `configure_time_options` call,
but that was not being called in all the right places.

This changes the setting of `use_time = true` into a
`configure_connection` hook instead.