quote_table_name fixes for 7.2 adapter changes
Rails changed the adapter contract in
0016280f4fde55d96738887093dc333aae0d107b such that `quote_table_name` as
an instance method gets mapped to `self.class.quote_table_name`, which
maps to `self.class.quote_column_name`, which is now default not
implemented. Since we override `quote_column_name` at the instance, we
can do the same to avoid breaking callers that rely on it.